Factorize uses of `isBrowser`, and move it from window-utils to api-utils/window/utils`.
