+++ packages/api-utils/lib/keyboard/observer.js
@@ -8,7 +8,8 @@
 const { Trait } = require("../light-traits");
 const { EventEmitterTrait: EventEmitter } = require("../events");
 const { DOMEventAssembler } = require("../events/assembler");
-const { browserWindowIterator, isBrowser } = require('../window-utils');
+const { browserWindowIterator } = require('../window-utils');
+const { isBrowser } = require('../window/utils');
 const { observer: windowObserver } = require("../windows/observer");
 
 // Event emitter objects used to register listeners and emit events on them
