+++ packages/api-utils/tests/test-window-observer.js
@@ -12,7 +12,8 @@ exports["test unload window observer"] = function(assert, done) {
   let loader = Loader(module);
 
   let utils = loader.require("api-utils/window-utils");
-  let { isBrowser, activeBrowserWindow: activeWindow } = utils;
+  let { activeBrowserWindow: activeWindow } = utils;
+  let { isBrowser } = require('api-utils/window/utils');
   let observer = loader.require("api-utils/windows/observer").observer;
   let opened = 0;
   let closed = 0;
